requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Django==5.1
django-nested-admin==4.1.1
gunicorn==23.0.0
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-monkey-business==1.1.0
pywin32-ctypes==0.2.3
setuptools==73.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1